modernize - complete small refactors (series can wait)
