+++ lib/concurrent.js
@@ -1,7 +1,7 @@
 'use strict';
 
 module.exports = {
-    concurrent: function(functions, callback) {
+    concurrent(functions, callback) {
         if (functions.length === 0) {
             return callback(null, []);
         }
