+++ lib/map-concurrent-all.js
@@ -3,12 +3,8 @@
 const concurrentAll = require('./concurrent-all.js').concurrentAll;
 
 module.exports = {
-    mapConcurrentAll: function(items, iteratee, callback) {
-
-        const toRun = items.map(function(item) {
-            return iteratee.bind(null, item);
-        });
-
+    mapConcurrentAll(items, iteratee, callback) {
+        const toRun = items.map((item) => iteratee.bind(null, item));
         concurrentAll(toRun, callback);
     }
 };
